se agrega metodod e insercion de forma correcta funcionando  de forma completa.

+++ DsiCodeTech.SuPlazaWeb.Contabilidad/Dto/ImpuestoDto.cs
@@ -10,7 +10,7 @@ namespace DsiCodeTech.SuPlazaWeb.Contabilidad.Dto
         public int id { get; set; }
         public string cod_barras { get; set; }
         public string descripcion { get; set; }
-        public double porcentaje { get; set; } 
+        public decimal porcentaje { get; set; } 
         public DateTime fecha_registro { get; set; }
     }
 }
